requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
jinja2==3.1.2
python-multipart==0.0.5
PyPDF2==3.0.1
Pillow==9.5.0
requests==2.28.2
